refactor: rename kebab-case module files to camelCase
