+++ src/components/overpay/OverpayComparisonChart.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import type { ChartConfig } from "@/components/ui/chart";
-import type { OverpayAnalysisResult } from "@/lib/loans/overpay-types";
+import type { OverpayAnalysisResult } from "@/lib/loans/overpayTypes";
 import { LazyChartBase as ChartBase } from "@/components/charts/LazyChartBase";
 import { currencyFormatter } from "@/constants";
 import { useShowPresentValue } from "@/hooks/useStoreSelectors";
