+++ src/components/overpay/OverpayVerdict.tsx
@@ -5,7 +5,7 @@ import {
   InformationCircleIcon,
 } from "@hugeicons/core-free-icons";
 import { HugeiconsIcon } from "@hugeicons/react";
-import type { RecommendationType } from "@/lib/loans/overpay-types";
+import type { RecommendationType } from "@/lib/loans/overpayTypes";
 import { Alert, AlertTitle, AlertDescription } from "@/components/ui/alert";
 
 const verdictConfig: Record<
