+++ src/lib/detailPages.ts
