+++ src/types/insightCards.ts
